Refactor modules for i18n and playground.
