--- core-kotox-i18n/build.gradle.kts
@@ -1,14 +1,0 @@
-plugins {
-    kotlin("jvm")
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
-}
-
-
-dependencies {
-
-    implementation(libs.libphonenumber)
-}
