+++ core-kotox/build.gradle.kts
@@ -6,3 +6,8 @@ java {
     sourceCompatibility = JavaVersion.VERSION_11
     targetCompatibility = JavaVersion.VERSION_11
 }
+
+dependencies {
+
+    implementation(libs.libphonenumber)
+}
