+++ core-kotox/src/main/kotlin/cz/kotox/android/core/domain/Mapper.kt
@@ -1,0 +1,5 @@
+package cz.kotox.core.domain
+
+interface Mapper<I, O> {
+    fun map(from: I): O
+}
