+++ kotox-i18n-data/proguard-rules.pro
