+++ kotox-i18n-data/src/main/kotlin/cz/kotox/data/api/repository/CountryRepository.kt
@@ -1,0 +1,9 @@
+package cz.kotox.data.api.repository
+
+import cz.kotox.data.api.model.CountryModel
+
+
+interface CountryRepository {
+    suspend fun getPhoneNumberPrefixModels(): List<CountryModel>
+    suspend fun getPhoneNumberPrefixCountryMap(): Map<String, Int>
+}
