+++ mobile-kotox-task/src/main/kotlin/cz/kotox/task/list/di/AppModule.kt
@@ -1,6 +1,8 @@
 package cz.kotox.task.list.di
 
 import cz.kotox.android.core.config.AppProperties
+import cz.kotox.core.network.config.AppNetworkingProperties
+import cz.kotox.task.list.config.KotoxTaskAppNetworkingProperties
 import cz.kotox.task.list.config.KotoxTaskAppProperties
 import dagger.Module
 import dagger.Provides
@@ -13,5 +15,9 @@ object AppModule {
 
     @Provides
     fun provideAppProperties(): AppProperties = KotoxTaskAppProperties()
+
+    @Provides
+    fun provideAppNetworkingProperties(): AppNetworkingProperties =
+        KotoxTaskAppNetworkingProperties()
 
 }
